Rename files so they refer to the service instead of to getting the HTML
